finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -2,7 +2,7 @@ monthly_income = int(input("Enter your monthly income: "))
 
 total_monthly_expenses = int(input("Enter your total monthly expenses: "))
 
-monthly_savings = int(monthly_income - total_monthly_expenses)
+monthly_savings = float(monthly_income) - float(total_monthly_expenses)
 
 project_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
